deployments/worker2/docker-img/requirements.txt
kafka-python
pandas

